@@ -67,7 +67,7 @@ abstract class OpenSSH
         // key format is described here:
         // https://cvsweb.openbsd.org/cgi-bin/cvsweb/src/usr.bin/ssh/PROTOCOL.key?annotate=HEAD
 
-        if (strpos($key, 'BEGIN OPENSSH PRIVATE KEY') !== false) {
+        if (str_contains($key, 'BEGIN OPENSSH PRIVATE KEY')) {
             $key = preg_replace('#(?:^-.*?-[\r\n]*$)|\s#ms', '', $key);
             $key = Base64::decode($key);
             $magic = Strings::shift($key, 15);
