@@ -103,7 +103,7 @@ abstract class PSS extends Progenitor
             throw new \UnexpectedValueException('Key should be a string - not a ' . gettype($key));
         }
 
-        $components = ['isPublicKey' => strpos($key, 'PUBLIC') !== false];
+        $components = ['isPublicKey' => str_contains($key, 'PUBLIC')];
 
         $key = parent::load($key, $password);
 
