@@ -18,7 +18,6 @@ use phpseclib3\Crypt\EC\BaseCurves\TwistedEdwards;
 use phpseclib3\Crypt\Hash;
 use phpseclib3\Crypt\Random;
 use phpseclib3\Math\BigInteger;
-use phpseclib3\Math\PrimeField\Integer;
 
 class Ed25519 extends TwistedEdwards
 {
@@ -182,7 +181,7 @@ class Ed25519 extends TwistedEdwards
 
         return [
             'dA' => $dA,
-            'secret' => $str
+            'secret' => $str,
         ];
     }
 
